requirements.txt
 # -----------------------------
# Core utilities
# -----------------------------
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.0
httpx>=0.27.0
tenacity>=8.3.0
tqdm>=4.66.0
rich>=13.8.0
loguru>=0.7.2

# -----------------------------
# Data handling (The "Kitchen")
# -----------------------------
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=16.0.0
polars>=1.5.0  # The parallel chef
openpyxl>=3.1.0 # Needed to read Excel files (common in banking)
pyyaml>=6.0.0

# -----------------------------
# The Lab (Notebooks & Visualization)
# -----------------------------
jupyterlab>=4.0.0
matplotlib>=3.8.0
seaborn>=0.13.0 # Makes matplotlib prettier
ipywidgets>=8.1.0 # Interactive notebook elements

# -----------------------------
# Hugging Face + datasets
# -----------------------------
datasets>=2.20.0
huggingface_hub>=0.24.0

# -----------------------------
# Kaggle downloads
# -----------------------------
kaggle>=1.6.17
kagglehub>=0.3.7

# -----------------------------
# AWS/S3
# -----------------------------
boto3>=1.34.0
s3fs>=2024.6.1
fsspec>=2024.6.1

# -----------------------------
# Documents / images
# -----------------------------
pillow>=10.4.0
pdf2image>=1.17.0
pymupdf>=1.24.0
opencv-python-headless>=4.10.0.0
pytesseract>=0.3.10
rapidfuzz>=3.9.0

# -----------------------------
# Local ML & Vector DB (The "Brain" & "Memory")
# -----------------------------
torch>=2.2.0
transformers>=4.39.0
einops>=0.7.0     # Required for ColPali/Vision models
timm>=0.9.0       # Required for Vision models
qdrant-client>=1.8.0 # The Vector Database

# -----------------------------
# API
# -----------------------------
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# -----------------------------
# Agentic framework
# -----------------------------
langchain>=0.2.12
langchain-community>=0.2.11
langgraph>=0.2.0
langchain-qdrant>=0.1.0 # Connector for Qdrant

# -----------------------------
# Eval / Observability
# -----------------------------
langsmith>=0.1.90
ragas>=0.1.14
mlflow>=2.15.0

# -----------------------------
# Dev tooling
# -----------------------------
pytest>=8.2.0
ruff>=0.5.7
black 
mypy 
pre-commit 
types-requests 
types-tqdm
rapidfuzz